requirements.txt
hikari-lightbulb
python-dotenv
ezgiphy